Changed to flat data structures. Unified BasicSet<Interval> classes (except Rectangle)

+++ include/declarations.h
@@ -34,7 +34,7 @@
 
 
 namespace Ariadne { namespace Base {
-template<class T, size_t N=0> class array;
+template<class T, unsigned short int N=0> class array;
 template<class T> class sequence;
 }}
